pyproject.toml
[tool.poetry]
name = "course-drf"
version = "0.1.0"
description = ""
authors = ["Skillova <skillova@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
django = "4.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.1"
django-celery-beat = "^2.6.0"
redis = "^5.0.8"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
